Desgaste double

+++ src/model/Desgaste.java
@@ -1,5 +1,5 @@
 package model;
 
 public abstract class Desgaste {
-	abstract int calcularDesgaste(int fuerza);
+	abstract double calcularDesgaste(int fuerza);
 }
